pyproject.toml
[build-system]
requires = ["setuptools", "cmake>=3.18"]
build-backend = "setuptools.build_meta"

[project]
name="mujoco-warp"
version = "0.0.6"
# TODO(team): create a distribution list
authors = [
    {name = "Newton Developers", email = "mujoco@deepmind.com"},
]
description = "MuJoCo Warp (MJWarp)"
readme = {file = "README.md", content-type = "text/markdown"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
requires-python = "==3.10.*"
dependencies = [
    "absl-py",
    "etils[epath]",
    "mujoco==3.4.0",
    "numpy",
    # "axinfra",
    "axinfra @ git+ssh://git@sh-code.mthreads.com/sw/axInfrastructure.git@0.0.4",
    # "setuptools",
]

[tool.uv]
managed = true

[[tool.uv.index]]
name = "nvidia"
url = "https://pypi.nvidia.com"
explicit = true

[[tool.uv.index]]
name = "mujoco"
url = "https://py.mujoco.org/"
explicit = true

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[tool.uv.sources]
# warp-lang = {index = "nvidia"}
# mujoco = {index = "mujoco"}
axinfra = { workspace = true }

[tool.uv.workspace]
members = [
    "dependencies/mujoco-musa/dependencies/axinfra",
]

[project.optional-dependencies]
dev = [
    "asv",
    "pre-commit",
    "pytest",
    "pytest-xdist",
    "ruff",
    "pygls>=1.0.0,<2.0.0",
    "lsprotocol>=2023.0.1,<2024.0.0",
]
warp = [
    "warp-lang>=1.9.1",
]
# TODO(team): cpu and cuda JAX optional dependencies are temporary, remove after we land MJX:Warp
cpu = [
    "jax"
]
cuda = [
    "jax[cuda12]"
]

[project.scripts]
mjwarp-testspeed = "mujoco_warp.testspeed:main"
mjwarp-viewer = "mujoco_warp.viewer:main"

[project.urls]
Homepage = "https://sh-code.mthreads.com/sw/mujoco_warp_musa"
Documentation = "https://mujoco.readthedocs.io"
Repository = "https://sh-code.mthreads.com/sw/mujoco_warp_musa"
Changelog = "https://mujoco.readthedocs.io/en/stable/changelog.html"

[tool.ruff]
line-length = 128
indent-width = 2
extend-exclude = ["*.ipynb"]

[tool.ruff.lint]
select = [
    "D",  # pydocstyle conventions and style
    "I",  # isort
    "W",  # pycodestyle
    "F401",  # unused imports
]

ignore = [
    "D100",  # missing docstring public module
    "D103",  # missing docstring public function
]

[tool.ruff.lint.isort]
force-single-line = true
single-line-exclusions = ["typing"]

[tool.ruff.lint.pycodestyle]
max-doc-length = 100
max-line-length = 128

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["D"]
"contrib/*" = ["D"]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 100

[tool.setuptools]
package-data = { "mujoco_warp" = ["test_data/*"], "mujoco_warp._src.mujoco_musa.bin" = ["*.so"] }

[tool.setuptools.packages.find]
include = ["mujoco_warp*"]
